bugs gone

+++ src/pages/pre-built/products/ProductList.js
@@ -22,15 +22,17 @@ import {
 import { Card, DropdownItem, UncontrolledDropdown, DropdownMenu, DropdownToggle } from "reactstrap";
 import { useForm } from "react-hook-form";
 import { Modal, ModalBody } from "reactstrap";
-import userDashBoard from "../../../zustand/userDashBoard/userDashBoard";
+import userDashBoard from "../../../zustand/DashBoard/userDashBoard";
+import { errorToast } from "../../../pages/components/misc/ReactToastify";
 import { Link } from "react-router-dom";
 
 const ProductList = () => {
-	const { data, getUserAllShortenLink, removeLink, deleteMany, updateLink } = userDashBoard((state) => ({
+	const { data, getUserAllShortenLink, removeLink, deleteMany, updateLink, getLinkById } = userDashBoard((state) => ({
 		data: state.allLinks,
 		getUserAllShortenLink: state.getUserAllShortenLink,
 		removeLink: state.removeLink,
 		deleteMany: state.deleteMany,
+		getLinkById: state.getLinkById,
 		updateLink: state.updateLink,
 	}));
 	useEffect(() => {
@@ -38,6 +40,7 @@ const ProductList = () => {
 	}, [getUserAllShortenLink]);
 	const [sm, updateSm] = useState(false);
 	const [formData, setFormData] = useState({
+		id: "",
 		protocol: "",
 		domain: "",
 		backlink: "",
@@ -46,7 +49,6 @@ const ProductList = () => {
 		official_domain: "shorterme.link/",
 		createdAt: "",
 	});
-	const [editId, setEditedId] = useState();
 	const [view, setView] = useState({
 		edit: false,
 		add: false,
@@ -73,12 +75,13 @@ const ProductList = () => {
 
 	// // function to close the form modal
 	const onFormCancel = () => {
-		setView({ edit: false, add: false, details: false });
 		resetForm();
+		setView({ edit: false, add: false, details: false });
 	};
 
 	const resetForm = () => {
 		setFormData({
+			id: "",
 			protocol: "",
 			domain: "",
 			backlink: "",
@@ -96,40 +99,27 @@ const ProductList = () => {
 			backlink: data.backlink,
 			domain: data.domain,
 		};
-		await updateLink(submit, editId);
+		await updateLink(submit, formData.id);
 		resetForm();
 
 		setView({ edit: false, add: false });
 	};
 
 	// function that loads the want to editted data
-	const onEditClick = (id) => {
-		data.forEach((item) => {
-			if (item.id === id) {
-				setFormData({
-					protocol: item.protocol,
-					domain: item.domain,
-					backlink: item.backlink,
-					shorterLink: item.shorterLink,
-					link_title: item.link_title,
-					official_domain: "shorterme.link/",
-					createdAt: new Date(item.createdAt).toLocaleString(),
-				});
-			}
+	const onEditClick = async (id) => {
+		const data = await getLinkById(id);
+		setFormData({
+			id: data.data.data.id,
+			protocol: data.data.data.protocol,
+			domain: data.data.data.domain,
+			backlink: data.data.data.backlink,
+			shorterLink: data.data.data.shorterLink,
+			link_title: data.data.data.link_title,
+			official_domain: "shorterme.link/",
+			createdAt: new Date(data.data.data.createdAt).toLocaleString(),
 		});
-		setEditedId(id);
 		setView({ add: false, edit: true });
 	};
-
-	// selects all the products
-	// const selectorCheck = (e) => {
-	// 	let newData;
-	// 	newData = data.map((item) => {
-	// 		item.check = e.currentTarget.checked;
-	// 		return item;
-	// 	});
-	// 	setData([...newData]);
-	// };
 
 	// selects one product
 	let linkId = [];
@@ -138,6 +128,9 @@ const ProductList = () => {
 	};
 	// function to delete the selected item
 	const selectorDeleteProduct = () => {
+		if (linkId.length === 0) {
+			return errorToast(`Please select the item before click ❌❌`, "Error");
+		}
 		deleteMany(linkId);
 		linkId = [];
 	};
@@ -233,17 +226,7 @@ const ProductList = () => {
 							<div className="card-inner p-0">
 								<DataTableBody>
 									<DataTableHead>
-										<DataTableRow className="nk-tb-col-check">
-											<div className="custom-control custom-control-sm custom-checkbox notext">
-												<input
-													type="checkbox"
-													className="custom-control-input form-control"
-													id="uid_1"
-													// onChange={(e) => selectorCheck(e)}
-												/>
-												<label className="custom-control-label" htmlFor="uid_1"></label>
-											</div>
-										</DataTableRow>
+										<DataTableRow className="nk-tb-col-check"></DataTableRow>
 										<DataTableRow sm>
 											<span>title</span>
 										</DataTableRow>
@@ -268,14 +251,7 @@ const ProductList = () => {
 														<DropdownMenu right>
 															<ul className="link-list-opt no-bdr">
 																<li>
-																	<DropdownItem
-																		tag="a"
-																		href="#remove"
-																		onClick={(ev) => {
-																			ev.preventDefault();
-																			selectorDeleteProduct();
-																		}}
-																	>
+																	<DropdownItem tag="a" href="#remove" onClick={selectorDeleteProduct}>
 																		<Icon name="trash"></Icon>
 																		<span>Remove Selected</span>
 																	</DropdownItem>
@@ -296,7 +272,7 @@ const ProductList = () => {
 																<input
 																	type="checkbox"
 																	className="custom-control-input form-control"
-																	defaultChecked={item.check}
+																	defaultChecked={false}
 																	id={item.id + "uid1"}
 																	key={Math.random()}
 																	onChange={(e) => onSelectChange(e, item.id)}
@@ -305,22 +281,28 @@ const ProductList = () => {
 															</div>
 														</DataTableRow>
 														<DataTableRow sm>
-															<Link to={`${process.env.PUBLIC_URL}/kyc-details-regular/${item.id}`}>
-																<span className="tb-product align-items-start d-flex flex-column">
-																	<span className="title d-lg-none">
-																		{item.link_title ? item.link_title.slice(0, 25) + "..." : "No title"}
-																	</span>
-																	<span className="title d-none  d-lg-block">
-																		{item.link_title ? item.link_title.slice(0, 60) + "..." : "No title"}
-																	</span>
-																	<span className="tb-sub d-lg-none">{`${item.protocol}://${item.domain}/${
-																		item.backlink.slice(0, 25) + "..."
-																	}`}</span>
-																	<span className="tb-sub d-none  d-lg-block">{`${item.protocol}://${item.domain}/${
-																		item.backlink.slice(0, 60) + "..."
-																	}`}</span>
-																</span>
-															</Link>
+															{/* <Link to={`${process.env.PUBLIC_URL}/kyc-details-regular/${item.id}`}> */}
+															<span className="tb-product align-items-start d-flex flex-column">
+																<Link
+																	to={`${process.env.PUBLIC_URL}/kyc-details-regular/${item.id}`}
+																	className="title d-lg-none"
+																>
+																	{item.link_title ? item.link_title.slice(0, 25) + "..." : "No title"}
+																</Link>
+																<Link
+																	to={`${process.env.PUBLIC_URL}/kyc-details-regular/${item.id}`}
+																	className="title d-none  d-lg-block"
+																>
+																	{item.link_title ? item.link_title.slice(0, 60) + "..." : "No title"}
+																</Link>
+																<span className="tb-sub d-lg-none">{`${item.protocol}://${item.domain}/${
+																	item.backlink.slice(0, 25) + "..."
+																}`}</span>
+																<span className="tb-sub d-none  d-lg-block">{`${item.protocol}://${item.domain}/${
+																	item.backlink.slice(0, 60) + "..."
+																}`}</span>
+															</span>
+															{/* </Link> */}
 														</DataTableRow>
 														<DataTableRow>
 															<a
@@ -356,7 +338,6 @@ const ProductList = () => {
 																						tag="a"
 																						href="#edit"
 																						onClick={(ev) => {
-																							ev.preventDefault();
 																							onEditClick(item.id);
 																							toggle("edit");
 																						}}
@@ -426,7 +407,7 @@ const ProductList = () => {
 						<div className="p-2">
 							<h5 className="title">Update Link</h5>
 							<div className="mt-4">
-								<form noValidate onSubmit={handleSubmit(onEditSubmit)}>
+								<form onSubmit={handleSubmit(onEditSubmit)}>
 									<Row className="g-3">
 										<Col size="12">
 											<div className="form-group">
